WeSee/object_models/object_detection_upload.py
#!/usr/bin/env python3
"""
Object Detection Upload Module - Simplified Version
Handles object detection for uploaded images using basic OpenCV
"""

from PIL import Image
import os
import sys
from datetime import datetime
from flask import Flask
import subprocess
import cv2
import numpy as np

# Add parent directory to path to avoid utils import conflict  
sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))

def detect_objects(file_path):
    """
    Simple object detection for uploaded images
    Returns basic detection results and audio description
    """
    print(f"Processing uploaded file: {file_path}")
    
    try:
        # Read the image
        image = cv2.imread(file_path)
        if image is None:
            print("Could not load image")
            return "Could not process the uploaded image", None, datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
        
        # Simple detection approach - just analyze image properties
        height, width, channels = image.shape
        
        # Basic analysis - this is a simplified version
        # In a real implementation, you'd use ONNX models here too
        detected_objects = ["uploaded_image"]  # Placeholder detection
        
        timestamp = datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
        
        # Generate description
        description_text = f"Successfully processed uploaded image. Image dimensions: {width}x{height} pixels."
        
        # Generate audio description
        audio_path = _generate_audio_description(description_text, timestamp)
        
        print(f"Upload detection completed: {description_text}")
        return description_text, audio_path, timestamp
        
    except Exception as e:
        print(f"Error processing upload: {e}")
        timestamp = datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
        error_msg = f"Error processing uploaded file: {str(e)}"
        audio_path = _generate_audio_description(error_msg, timestamp)
        return error_msg, audio_path, timestamp

def _generate_audio_description(description_text, timestamp):
    """Generate audio file from description text using Windows TTS"""
    try:
        from flask import current_app
        audio_folder = current_app.config.get('AUDIO_FOLDER', 'static/audio')
    except:
        audio_folder = 'static/audio'
    
    audio_path = os.path.join(audio_folder, f'description_{timestamp}.wav')
    os.makedirs(audio_folder, exist_ok=True)
    
    # Use Windows Speech API for TTS
    try:
        # Escape quotes in the description text
        safe_text = description_text.replace('"', '').replace("'", "")
        
        subprocess.run([
            "powershell", "-Command",
            f"Add-Type -AssemblyName System.Speech; "
            f"$synth = New-Object System.Speech.Synthesis.SpeechSynthesizer; "
            f"$synth.SetOutputToWaveFile('{audio_path}'); "
            f"$synth.Speak('{safe_text}'); "
            f"$synth.Dispose()"
        ], check=True, capture_output=True, timeout=30)
        
        print(f"Audio generated: {audio_path}")
        
    except Exception as e:
        print(f"TTS Error: {e}")
        # Create a silent audio file as fallback
        try:
            import wave
            with wave.open(audio_path, 'w') as wav_file:
                wav_file.setnchannels(1)
                wav_file.setsampwidth(2)
                wav_file.setframerate(16000)
                # Write 1 second of silence
                wav_file.writeframes(b'\x00' * 32000)
        except Exception as wav_error:
            print(f"Could not create fallback audio: {wav_error}")
    
    return audio_path

# For compatibility with main.py
def object_detection_upload():
    """Compatibility function"""
    return detect_objects